pattern.test.ts 삭제, validation 테스트 코드로 분리 validation만 테스트하는 파일로 수정함
